pyproject.toml
[project]
name = "farmlanding"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "django>=5.2.2",
    "django-phonenumber-field>=8.1.0",
    "phonenumbers>=9.0.6",
    "pillow>=11.2.1",
    "python-dotenv>=1.1.0",
    "requests>=2.32.3",
]
